--- compiler-plugin/src/main/kotlin/arrow/meta/Test.kt
@@ -1,6 +1,0 @@
-package arrow.meta
-
-
-fun main() {
-  println("jsss")
-}
