+++ compiler-plugin/src/main/kotlin/arrow/meta/smt/dsl/scope/array/ArrayCtx.kt
