+++ compiler-plugin/src/main/kotlin/arrow/meta/smt/dsl/scope/quantify/QuantifyCtx.kt
