+++ compiler-plugin/src/main/kotlin/arrow/meta/smt/dsl/scope/slogic/SLogicCtx.kt
