+++ compiler-plugin/src/main/kotlin/arrow/meta/smt/dsl/scope/vector/VectorCtx.kt
