+++ compiler-plugin/src/main/kotlin/arrow/meta/smt/dsl/solver/SolverAlgebra.kt
