+++ compiler-plugin/src/main/kotlin/arrow/meta/smt/dsl/solver/cvc4/Cvc4Scope.kt
