--- meta-smt/build.gradle
@@ -1,13 +1,0 @@
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$KOTLIN_VERSION"
-    compile "org.sosy-lab:java-smt:$JAVA_SMT"
-    compile "org.sosy-lab:javasmt-solver-z3:$JAVASMT_SOLVER_Z3"
-}
-
-
-compileKotlin {
-    kotlinOptions.jvmTarget = "$JVM_TARGET_VERSION"
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "$JVM_TARGET_VERSION"
-}
