+++ settings.gradle
@@ -4,4 +4,3 @@ include ':idea-plugin'
 include ':meta-test'
 include ':docs'
 include ':prelude'
-include 'meta-smt'
